添加注释

+++ app/Http/Controllers/UsersController.php
@@ -7,28 +7,50 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
 
+/**
+ * 用户管理
+ *
+ * Class UsersController
+ * @package App\Http\Controllers
+ */
 class UsersController extends Controller
 {
 
     public function __construct()
     {
+        //中间件
         $this->middleware('auth', [
             'except' => ['show', 'create', 'store', 'index', 'confirmEmail']
         ]);
     }
 
+    /**
+     * 用户列表
+     *
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View\
+     */
     public function index()
     {
         $users = User::paginate(10);
         return view("users.index", compact('users'));
     }
 
-
+    /**
+     * 创建用户
+     *
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function create()
     {
         return view("users.create");
     }
 
+    /**
+     * 用户个人中心
+     *
+     * @param User $user
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function show(User $user)
     {
         $statuses = $user->statuses()
@@ -37,43 +59,73 @@ class UsersController extends Controller
         return view('users.show', compact('user', 'statuses'));
     }
 
-
+    /**
+     * 创建用户
+     *
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Validation\ValidationException
+     */
     public function store(Request $request)
     {
+        //验证字段
         $this->validate($request, [
             'name' => 'required|max:50',
             'email' => 'required|email|unique:users|max:255',
             'password' => 'required|confirmed|min:6'
         ]);
+        //创建用户
         $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
             'password' => bcrypt($request->password),
         ]);
-
+        //发送邮件
         $this->sendEmailConfirmationTo($user);
         session()->flash('success', '验证邮件已发送到你的注册邮箱上，请注意查收。');
 
         return redirect()->route('users.show', [$user]);
     }
 
+    /**
+     * 编辑用户
+     *
+     * @param User $user
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function edit(User $user)
     {
+        //验证权限
         $this->authorize('update', $user);
         return view('users.edit', compact('user'));
     }
 
+    /**
+     * 更新用户
+     *
+     * @param User $user
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @throws \Illuminate\Validation\ValidationException
+     */
     public function update(User $user, Request $request)
     {
+        //验证权限
         $this->authorize('update', $user);
+        //验证字段
         $this->validate($request, [
             'name' => 'required|max:50',
             'password' => 'nullable|confirmed|min:6'
         ]);
+        //用户名称
         $data['name'] = $request->name;
+        //是否修改密码
         if ($request->password) {
             $data['password'] = bcrypt($request->password);
         }
+        //更新数据
         $user->update($data);
 
         session()->flash('success', '个人资料更新成功！');
@@ -81,15 +133,29 @@ class UsersController extends Controller
         return redirect()->route('users.show', $user->id);
     }
 
-
+    /**
+     * 删除用户
+     *
+     * @param User $user
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Exception
+     */
     public function destroy(User $user)
     {
+        //验证权限
+        $this->authorize('destroy', $user);
+        //删除用户
         $user->delete();
         session()->flash('success', '成功删除用户！');
         return back();
     }
 
-
+    /**
+     * 关注页面
+     *
+     * @param User $user
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function followings(User $user)
     {
         $users = $user->followings()->paginate(30);
@@ -97,7 +163,12 @@ class UsersController extends Controller
         return view('users.show_follow', compact('users', 'title'));
     }
 
-
+    /**
+     * 粉丝页面
+     *
+     * @param User $user
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function followers(User $user)
     {
         $users = $user->followers()->paginate(30);
@@ -105,7 +176,27 @@ class UsersController extends Controller
         return view('users.show_follow', compact('users', 'title'));
     }
 
-
+    /**
+     * 激活邮件
+     *
+     * @param $token
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function confirmEmail($token)
+    {
+        $user = User::where('activation_token', $token)->firstOrFail();
+        $user->activated = true;
+        $user->activation_token = null;
+        $user->save();
+        Auth::login($user);
+        session()->flash('success', '恭喜你，激活成功！');
+        return redirect()->route('users.show', [$user]);
+    }
+    /**
+     * 发送验证邮件
+     *
+     * @param $user
+     */
     protected function sendEmailConfirmationTo($user)
     {
         $view = 'emails.confirm';
@@ -120,14 +211,5 @@ class UsersController extends Controller
     }
 
 
-    public function confirmEmail($token)
-    {
-        $user = User::where('activation_token', $token)->firstOrFail();
-        $user->activated = true;
-        $user->activation_token = null;
-        $user->save();
-        Auth::login($user);
-        session()->flash('success', '恭喜你，激活成功！');
-        return redirect()->route('users.show', [$user]);
-    }
+
 }
